Searching for friends functionality added

+++ InteractiveChat/Controllers/UserController.cs
@@ -1,6 +1,7 @@
-﻿using InteractiveChat.Data;
+﻿using AutoMapper;
+using InteractiveChat.Data;
+using InteractiveChat.DTOs;
 using InteractiveChat.Models;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
@@ -12,40 +13,56 @@ namespace InteractiveChat.Controllers
         private readonly ApplicationDbContext _context;
         private readonly UserManager<ApplicationUser> _userManager;
         private readonly IWebHostEnvironment _webHostEnvironment;
+        private readonly IMapper _mapper;
 
         public UserController(
             ILogger<HomeController> logger,
-            ApplicationDbContext context, 
+            ApplicationDbContext context,
             UserManager<ApplicationUser> userManager,
-            IWebHostEnvironment webHostEnvironment)
+            IWebHostEnvironment webHostEnvironment,
+            IMapper mapper
+            )
+
         {
             _logger = logger;
             _context = context;
             _userManager = userManager;
             _webHostEnvironment = webHostEnvironment;
+            _mapper = mapper;
         }
         public IActionResult Index()
         {
             return View();
         }
+        [HttpGet]
         public async Task<IActionResult> EditAsync()
         {
             ApplicationUser user = await _userManager.GetUserAsync(HttpContext.User);
-            return View(user);
+            var userDto = _mapper.Map<ApplicationUserDTO>(user);
+            return View(userDto);
         }
+
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(ApplicationUser applicationUser)
+        public async Task<IActionResult> Edit(ApplicationUserDTO applicationUserDto)
         {
             if (ModelState.IsValid)
             {
                 var user = await _userManager.GetUserAsync(HttpContext.User);
-                if (user == null) { 
+                if (user == null)
+                {
                     return NotFound();
                 }
-                user.FirstName = applicationUser.FirstName;
-                user.LastName = applicationUser.LastName;
-                user.ProfilePicUrl = SaveProfilePicAsync(applicationUser);
+
+                // Update only the necessary properties
+                user.FirstName = applicationUserDto.FirstName;
+                user.LastName = applicationUserDto.LastName;
+
+                // Handle the profile picture
+                if (applicationUserDto.ProfilePicUrl != null)
+                {
+                    user.ProfilePicUrl = SaveProfilePicAsync(applicationUserDto.ProfilePicUrl);
+                }
                 var result = await _userManager.UpdateAsync(user);
                 if (result.Succeeded)
                 {
@@ -55,12 +72,12 @@ namespace InteractiveChat.Controllers
                 {
                     ModelState.AddModelError(string.Empty, error.Description);
                 }
-               
+
             }
-            return View(applicationUser);
+            return View(applicationUserDto);
         }
 
-        private string SaveProfilePicAsync(ApplicationUser user)
+        private string SaveProfilePicAsync(string userProfilePic)
         {
             var profilePicsDirectory = Path.Combine(_webHostEnvironment.WebRootPath, @"images\users");
             var profilePicFile = Request.Form.Files["ProfilePicUrl"];
@@ -74,8 +91,8 @@ namespace InteractiveChat.Controllers
                 }
 
                 // Delete old profile picture if exists
-                string oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, user.ProfilePicUrl.TrimStart('\\'));
-                if (System.IO.File.Exists(oldImagePath))
+                string oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, userProfilePic.TrimStart('\\'));
+                if (System.IO.File.Exists(oldImagePath) && !oldImagePath.Contains("profile-placeholder.png"))
                 {
                     System.IO.File.Delete(oldImagePath);
                 }
@@ -90,6 +107,10 @@ namespace InteractiveChat.Controllers
                     profilePicFile.CopyTo(fileStream);
                 }
 
+            }
+            else
+            {
+                uniqueFileName = "profile-placeholder.png";
             }
             // Save file path to database/users table
             return @"\images\users\" + uniqueFileName;
